Avoid RMW for 24-bit Z writes.
Fixes an aliasing problem where color is 32-bit and Z is 24-bit, where
color and Z will race, but only the alpha channel for color is actually
considered important. This race can be resolved if we use u8vec3 writes
instead. Also avoids latency to RMW memory when writing Z24, which is
nice I suppose.