Immediately invalidate texture cache after a transfer operation.
Potential failure here was having a texture cached already,
then LocalCopy would write to texture, and then sampled from that
address.

In this case, we will not have observed an end-of-render-pass situation,
which was the only place the texture cache would be flushed.

Immediately flushing on transfer fixes this problem.