Be a bit smarter how we handle FBMSK and opacity.
If all FBMSK values match, a pixel can still be considered opaque,
since there is no risk of previous pixel values leaking through.

Also, improve the heuristic for when to decay SPRITE only passes to
single-sampled. Z testing should be allowed in these passes, as long as
there are no Z writes in the pass.